joined routes test in one index relates #15
