+++ __test__/routes/index.test.js
@@ -1,0 +1,9 @@
+const loginTests = require('./login');
+const signupTests = require('./signup');
+
+const runRoutesTests = async () => {
+  await signupTests();
+  await loginTests();
+};
+
+runRoutesTests();
